pyproject.toml
[project]
name = "facegram"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.129.0",
    "fastapi-users-db-sqlalchemy>=7.0.0",
    "fastapi-users[sqlalchemy]>=15.0.4",
    "imagekitio[aiohttp,standard]>=5.2.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.40.0",
]
